Use a single project, rename files using partial classes
