--- src/SourceBuild/content/eng/tools/BinaryToolKit/BinaryToolCli/BinaryToolCli.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>$(NetCurrent)</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <AllowPreReleaseDependencies>true</AllowPreReleaseDependencies>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\BinaryTool\BinaryTool.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.CommandLine" Version="2.0.0-*" />
-  </ItemGroup>
-
-</Project>
