--- src/SourceBuild/content/eng/tools/BinaryToolKit/BinaryToolCli/NuGet.config
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <packageSources>
-    <clear />
-    <add key="artifactsPath" value="%ARTIFACTS_PATH%" />
-  </packageSources>
-</configuration>
